Progress on implementing new bluez peripheral struct
